:recycle: Merge all fixture providers into a single provider
