+++ src/fixture_providers/index.ts
@@ -1,5 +1,2 @@
-export * from './fixture_provider_consumer_api';
-export * from './fixture_provider_deployment_api';
-export * from './fixture_provider_management_api';
-export * from './fixture_provider_process_engine';
 export * from './process_instance_handler';
+export * from './test_fixture_provider';
