+++ test/1_process_engine/boundary_event_message_test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const should = require('should');
-const TestFixtureProvider = require('../../dist/commonjs').FixtureProviderProcessEngine;
+const TestFixtureProvider = require('../../dist/commonjs').TestFixtureProvider;
 
 describe('Message Boundary Event - ', () => {
