+++ test/1_process_engine/parse_bpmn_model_to_object_model.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const should = require('should');
-const TestFixtureProvider = require('../../dist/commonjs').FixtureProviderProcessEngine;
+const TestFixtureProvider = require('../../dist/commonjs').TestFixtureProvider;
 
 describe('Process-Engine   Parse BPMN Process into new object model', () => {
