+++ test/3_consumer_api/correlation/get_correlation_results.js
@@ -5,7 +5,7 @@ const uuid = require('uuid');
 
 const startCallbackType = require('@process-engine/consumer_api_contracts').StartCallbackType;
 
-const TestFixtureProvider = require('../../../dist/commonjs').FixtureProviderConsumerApi;
+const TestFixtureProvider = require('../../../dist/commonjs').TestFixtureProvider;
 
 describe('Consumer API:   GET  ->  /correlations/:correlation_id/process_models/:process_model_id/results', () => {
