+++ test/3_consumer_api/events/trigger_event.js
@@ -2,7 +2,7 @@
 
 const should = require('should');
 
-const TestFixtureProvider = require('../../../dist/commonjs').FixtureProviderConsumerApi;
+const TestFixtureProvider = require('../../../dist/commonjs').TestFixtureProvider;
 
 describe('Consumer API: POST  ->  /process_models/:process_model_id/correlations/:correlation_id/events/:event_id/trigger', () => {
