+++ test/3_consumer_api/get_process_models/get_process_model_by_id.js
@@ -2,7 +2,7 @@
 
 const should = require('should');
 
-const TestFixtureProvider = require('../../../dist/commonjs').FixtureProviderConsumerApi;
+const TestFixtureProvider = require('../../../dist/commonjs').TestFixtureProvider;
 
 describe('Consumer API:   GET  ->  /process_models/:process_model_id', () => {
