+++ test/3_consumer_api/user_tasks/get_process_model_user_tasks.js
@@ -2,7 +2,7 @@
 
 const should = require('should');
 
-const TestFixtureProvider = require('../../../dist/commonjs').FixtureProviderConsumerApi;
+const TestFixtureProvider = require('../../../dist/commonjs').TestFixtureProvider;
 const ProcessInstanceHandler = require('../../../dist/commonjs').ProcessInstanceHandler;
 
 describe('Consumer API:   GET  ->  /process_models/:process_model_id/userTasks', () => {
