+++ test/4_deployment_api/import_from_file.js
@@ -4,7 +4,7 @@ const should = require('should');
 const path = require('path');
 const uuid = require('uuid');
 
-const TestFixtureProvider = require('../../dist/commonjs').FixtureProviderDeploymentApi;
+const TestFixtureProvider = require('../../dist/commonjs').TestFixtureProvider;
 
 describe('Deployment API -> importBpmnFromFile', () => {
 
@@ -19,8 +19,8 @@ describe('Deployment API -> importBpmnFromFile', () => {
     testFixtureProvider = new TestFixtureProvider();
     await testFixtureProvider.initializeAndStart();
 
-    deploymentContextDefault = testFixtureProvider.context;
-    deploymentContextForbidden = testFixtureProvider.contextForbidden;
+    deploymentContextDefault = testFixtureProvider.context.defaultUser;
+    deploymentContextForbidden = testFixtureProvider.context.restrictedUser;
 
     const bpmnFolderLocation = testFixtureProvider.getBpmnDirectoryPath();
     processModelPath = path.join(bpmnFolderLocation, `${processModelId}.bpmn`);
@@ -101,11 +101,9 @@ describe('Deployment API -> importBpmnFromFile', () => {
 
   async function assertThatImportWasSuccessful() {
 
-    const executionContextFacade = await testFixtureProvider.createExecutionContextFacadeForContext(deploymentContextDefault);
-
     const processModelService = await testFixtureProvider.resolveAsync('ProcessModelService');
 
-    const existingProcessModel = await processModelService.getProcessModelById(executionContextFacade, processModelId);
+    const existingProcessModel = await processModelService.getProcessModelById(testFixtureProvider.executionContextFacade, processModelId);
 
     should.exist(existingProcessModel);
   }
